pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clawvatar-engine"
version = "0.1.0"
description = "Real-time 3D avatar animation engine — audio in, lip-synced video out. CPU-friendly, open source, built for OpenClaw."
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [
    {name = "Ruh AI", email = "vaibhav@ruh.ai"},
]
keywords = ["avatar", "lip-sync", "3d", "real-time", "openclaw", "video-conferencing", "blend-shapes"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "websockets>=12.0",
    "numpy>=1.26.0",
    "opencv-python-headless>=4.9.0",
    "moderngl>=5.10.0",
    "pillow>=10.2.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0.1",
    "python-multipart>=0.0.9",
    "torch>=2.2.0",
    "torchaudio>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.scripts]
clawvatar = "clawvatar.cli:main"

[project.urls]
Homepage = "https://github.com/ruh-ai/clawvatar-engine"
Repository = "https://github.com/ruh-ai/clawvatar-engine"
Issues = "https://github.com/ruh-ai/clawvatar-engine/issues"

[tool.setuptools.packages.find]
include = ["clawvatar*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
